requirements.txt
# VinoGen-CyberCore Requirements
# Python 3.12.7 Compatible

# Core Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning & Neural Networks
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
kaleido>=0.2.1

# Graph & Network Visualization
networkx>=3.1
pygraphviz>=1.11

# Image & Animation
imageio>=2.31.0
pillow>=9.5.0
imageio-ffmpeg>=0.4.8

# Terminal UI & Aesthetics
rich>=13.3.0
colorama>=0.4.6
tqdm>=4.65.0
tabulate>=0.9.0
pyfiglet>=0.8.0

# Utilities
pyyaml>=6.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
